+++ packages/storefront/src/components/Card.vue
@@ -59,7 +59,7 @@
 </script>
 
 <style scoped lang="scss">
-  @import '@porsche-design-system/components-js/styles';
+  @use '@porsche-design-system/components-js/styles' as *;
 
   $border-radius: 12px;