+++ packages/storefront/src/components/Disclaimer.vue
@@ -15,7 +15,7 @@
 </script>
 
 <style scoped lang="scss">
-  @import '@porsche-design-system/components-js/styles';
+  @use '@porsche-design-system/components-js/styles' as *;
 
   .disclaimer {
     padding: $pds-spacing-static-small $pds-spacing-static-large;