+++ packages/storefront/src/components/Menu.vue
@@ -27,7 +27,7 @@
 </script>
 
 <style scoped lang="scss">
-  @import '@porsche-design-system/components-js/styles';
+  @use '@porsche-design-system/components-js/styles' as *;
 
   .menu {
     width: 2.5rem;