+++ packages/storefront/src/components/PorscheFooter.vue
@@ -73,7 +73,7 @@
 </script>
 
 <style scoped lang="scss">
-  @import '@porsche-design-system/components-js/styles';
+  @use '@porsche-design-system/components-js/styles' as *;
 
   .main-footer {
     @include pds-grid;