+++ packages/storefront/src/components/PorscheHeader.vue
@@ -95,7 +95,7 @@
 </script>
 
 <style scoped lang="scss">
-  @import '@porsche-design-system/components-js/styles';
+  @use '@porsche-design-system/components-js/styles' as *;
 
   .main-nav {
     @include pds-grid;