+++ packages/storefront/src/components/PorscheMainBar.vue
@@ -21,7 +21,7 @@
 </script>
 
 <style scoped lang="scss">
-  @import '@porsche-design-system/components-js/styles';
+  @use '@porsche-design-system/components-js/styles' as *;
 
   .main-bar {
     @include pds-grid;