+++ packages/storefront/src/components/Search.vue
@@ -108,7 +108,7 @@
 </script>
 
 <style scoped lang="scss">
-  @import '@porsche-design-system/components-js/styles';
+  @use '@porsche-design-system/components-js/styles' as *;
 
   ul,
   li {