+++ packages/storefront/src/components/Sidebar.vue
@@ -97,7 +97,7 @@
 </script>
 
 <style scoped lang="scss">
-  @import '@porsche-design-system/components-js/styles';
+  @use '@porsche-design-system/components-js/styles' as *;
 
   ul,
   li {