+++ packages/storefront/src/pages/ComponentMetaOverview.vue
@@ -124,7 +124,7 @@
 </script>
 
 <style scoped lang="scss">
-  @import '@porsche-design-system/components-js/styles';
+  @use '@porsche-design-system/components-js/styles' as *;
 
   p-table {
     max-width: none !important;