+++ packages/storefront/src/pages/patterns/apps/example-porsche-com.vue
@@ -588,7 +588,7 @@
 </script>
 
 <style lang="scss" scoped>
-  @import '@porsche-design-system/components-js/styles';
+  @use '@porsche-design-system/components-js/styles' as *;
 
   .teaser {
     @include pds-grid;